move base performance detector handler to perf dir
